Various cleanups

+++ src/daemon/stack.c
@@ -1,8 +1,7 @@
 /* -*- Mode: C; tab-width: 8; indent-tabs-mode: nil; c-basic-offset: 8 -*-
  *
- * stack.c - Notification stack groups.
- *
  * Copyright (C) 2006 Christian Hammond <chipx86@chipx86.com>
+ * Copyright (C) 2010 Red Hat, Inc.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -19,6 +18,7 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA
  * 02111-1307, USA.
  */
+
 #include "config.h"
 #include "engines.h"
 #include "stack.h"
@@ -197,7 +197,7 @@ notify_stack_new (NotifyDaemon       *daemon,
 
         g_assert (daemon != NULL);
         g_assert (screen != NULL && GDK_IS_SCREEN (screen));
-        g_assert (monitor < gdk_screen_get_n_monitors (screen));
+        g_assert (monitor < (guint)gdk_screen_get_n_monitors (screen));
         g_assert (location != NOTIFY_STACK_LOCATION_UNKNOWN);
 
         stack = g_new0 (NotifyStack, 1);
